# Conflicts:
#	src/main/java/site/benepay/common/exception/GlobalExceptionHandler.java
#	src/main/java/site/benepay/domain/merchant/dto/MerchantCategoryResponseDto.java
#	src/main/java/site/benepay/domain/merchant/dto/MerchantRequestDto.java
#	src/main/java/site/benepay/domain/merchant/dto/MerchantResponseDto.java
#	src/main/java/site/benepay/domain/merchant/service/MerchantServiceImpl.java
#	src/main/java/site/benepay/domain/merchant/vo/Merchant.java
#	src/main/java/site/benepay/domain/merchant/vo/MerchantCategory.java
#	src/main/resources/mybatis/mapper/MerchantCategoryMapper.xml
#	src/main/resources/mybatis/mapper/MerchantMapper.xml

+++ src/main/java/site/benepay/domain/card/vo/CardMonthlyStatusVO.java
@@ -1,18 +1,17 @@
 package site.benepay.domain.card.vo;
 
+import java.time.LocalDateTime;
+
 import lombok.Getter;
 import lombok.Setter;
-
-import java.math.BigDecimal;
-import java.time.LocalDateTime;
 
 @Getter
 @Setter
 public class CardMonthlyStatusVO {
 
-    private Long cardMonthlyStatusId;
-    private Long userCardId;
-    private String targetYearMonth;
-    private BigDecimal totalSpendingAmount;
-    private LocalDateTime updatedAt;
+	private Long cardMonthlyStatusId;
+	private Long userCardId;
+	private String targetYearMonth;
+	private Long totalSpendingAmount;
+	private LocalDateTime updatedAt;
 }
